Add test for formatting of single notebook
Copy the unformatted spaces notebook and first check that the copy and
the source are the same. Then run blackbook's main with the path being
the full path to spaces.ipynb. Then check that the contents of the
formatted spaces.ipynb differs from the copied notebook, which has been
untouched by blackbook. Also delete the copied notebook to make test
atomic.